+++ src/lib/djerba/plugins/supplement/body/test/plugin_test.py
@@ -15,76 +15,76 @@ class TestSupplementaryPluginBody(PluginTester):
         self.maxDiff = None
         self.tmp = tempfile.TemporaryDirectory(prefix='djerba_')
         self.tmp_dir = self.tmp.name
-        sup_dir_var = 'DJERBA_TEST_DATA'
+        sup_dir_var = 'DJERBA_TEST_DIR'
         self.sup_dir = os.environ.get(sup_dir_var)
 
     def testPwgsSupplementary(self):
         test_source_dir = os.path.realpath(os.path.dirname(__file__))
-        json_location = os.path.join(self.sup_dir ,"supplement-plugin/report_json/pwgs.supplement.json")
+        json_location = os.path.join(self.sup_dir ,"plugins/supplement/report_json/pwgs.supplement.json")
         params = {
             self.INI: 'PWGS.supp.ini',
             self.JSON: json_location,
-            self.MD5: 'bba6863fa170828c95eb97c9e5c2c22e'
+            self.MD5: '9980b0f7f92c6e51f6928ce852a33b0d'
         }
         self.run_basic_test(test_source_dir, params)
 
     def testTarSupplementary(self):
         test_source_dir = os.path.realpath(os.path.dirname(__file__))
-        json_location = os.path.join(self.sup_dir ,"supplement-plugin/report_json/tar.supplement.json")
+        json_location = os.path.join(self.sup_dir ,"plugins/supplement/report_json/tar.supplement.json")
         params = {
             self.INI: 'TAR.supp.ini',
             self.JSON: json_location,
-            self.MD5: 'd3ba5a11f020af46392fab8bbf553bc2'
+            self.MD5: '7e899cd598fcf52ba126ba91a1d5c234'
         }
         self.run_basic_test(test_source_dir, params)
    
     def testTarFailSupplementary(self):
         test_source_dir = os.path.realpath(os.path.dirname(__file__))
-        json_location = os.path.join(self.sup_dir ,"supplement-plugin/report_json/tar.fail.supplement.json")
+        json_location = os.path.join(self.sup_dir ,"plugins/supplement/report_json/tar.fail.supplement.json")
         params = {
             self.INI: 'TAR.FAIL.supp.ini',
             self.JSON: json_location,
-            self.MD5: '2724e816e516614f7921328e763921f5'
+            self.MD5: 'f42de9c16acee44016cef3cbbf4b4d94'
         }
         self.run_basic_test(test_source_dir, params)
 
     def testWgtsSupplementary(self):
         test_source_dir = os.path.realpath(os.path.dirname(__file__))
-        json_location = os.path.join(self.sup_dir ,"supplement-plugin/report_json/wgts.supplement.json")
+        json_location = os.path.join(self.sup_dir ,"plugins/supplement/report_json/wgts.supplement.json")
         params = {
             self.INI: 'WGTS.supp.ini',
             self.JSON: json_location,
-            self.MD5: '0ccae7e250b8940c822c96f6f610ec7b'
+            self.MD5: '189fb556a113e51d3c9eb52d272892ae'
         }
         self.run_basic_test(test_source_dir, params)
 
     def testWgtsFailSupplementary(self):
         test_source_dir = os.path.realpath(os.path.dirname(__file__))
-        json_location = os.path.join(self.sup_dir ,"supplement-plugin/report_json/wgts.fail.supplement.json")
+        json_location = os.path.join(self.sup_dir ,"plugins/supplement/report_json/wgts.fail.supplement.json")
         params = {
             self.INI: 'WGTS.FAIL.supp.ini',
             self.JSON: json_location,
-            self.MD5: '512e46eb1915db5f8a93fdd047147922'
+            self.MD5: '94fddd6ebd7456294457ddcafa4e084a'
         }
         self.run_basic_test(test_source_dir, params)
 
     def testWgts40XSupplementary(self):
         test_source_dir = os.path.realpath(os.path.dirname(__file__))
-        json_location = os.path.join(self.sup_dir ,"supplement-plugin/report_json/wgts40X.supplement.json")
+        json_location = os.path.join(self.sup_dir ,"plugins/supplement/report_json/wgts40X.supplement.json")
         params = {
             self.INI: 'WGTS40X.supp.ini',
             self.JSON: json_location,
-            self.MD5: 'e14d2c46282840f3c64dd1ae1f9b1b44'
+            self.MD5: '3b60dd5df4836f56b8255ae9c69706aa'
         }
         self.run_basic_test(test_source_dir, params)
 
     def testWgts40XFailSupplementary(self):
         test_source_dir = os.path.realpath(os.path.dirname(__file__))
-        json_location = os.path.join(self.sup_dir ,"supplement-plugin/report_json/wgts40X.fail.supplement.json")
+        json_location = os.path.join(self.sup_dir ,"plugins/supplement/report_json/wgts40X.fail.supplement.json")
         params = {
             self.INI: 'WGTS40X.FAIL.supp.ini',
             self.JSON: json_location,
-            self.MD5: '719ff62201b0e1dde4c907260bb9062c'
+            self.MD5: '5f8209c64606b24855295a61bb7f52ed'
         }
         self.run_basic_test(test_source_dir, params)